CTBD_MULTI/requirements.txt
python-socketio
uvicorn
fastapi